+++ phpdbg_bp.h
@@ -26,7 +26,7 @@
 typedef struct _phpdbg_breakfile_t {
 	const char *filename;
 	long line;
-	size_t id;
+	int id;
 } phpdbg_breakfile_t;
 
 /**
@@ -35,7 +35,7 @@ typedef struct _phpdbg_breakfile_t {
 typedef struct _phpdbg_breaksymbol_t {
 	const char *symbol;
 	long opline_num;
-	size_t id;
+	int id;
 } phpdbg_breaksymbol_t;
 
 void phpdbg_set_breakpoint_file(const char*, const char* TSRMLS_DC);
